Added GetProfile to retrieve profile of users

+++ app/src/main/java/ayush/ggv/instau/presentation/components/CommentListItem.kt
@@ -32,7 +32,7 @@ import ayush.ggv.instau.ui.theme.SocialAppTheme
 fun CommentListItem(
     modifier: Modifier = Modifier,
     comment: Comment,
-    onProfileClick: (Int) -> Unit,
+    onProfileClick: (Long) -> Unit,
     onMoreIconClick: () -> Unit
 ) {
 
@@ -46,7 +46,7 @@ fun CommentListItem(
                 imageUrl = comment.authorImageUrl,
                 modifier = modifier.size(30.dp)
             ) {
-                onProfileClick(comment.authorId)
+                onProfileClick(comment.authorId.toLong())
             }
